requirements.txt
requests
PyMuPDF
phidata
django-cors-headers